// resources/js/game/components/kingdoms/capital-city/enums/queue-status.ts
export enum QueueStatus {
    TRAVELING = "traveling",

    PROCESSING = "processing",

    REQUESTING = "requesting",

    BUILDING = "building",

    REPAIRING = "repairing",

    RECRUITING = "recruiting",

    REJECTED = "rejected",

    FINISHED = "finished",

    CANCELLED = "cancelled",

    CANCELLATION_REJECTED = "cancellation_rejected",
}